+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Updated voiceover and translation credits for the Intro sample scene.
 - Added shadow line support to BuiltInLineProvider.
+- Added support for generating C# variable storage classes that expose properties for string, number and boolean variables found in a Yarn Project.
 
 ### Changed